garbage_detection/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.4
djangorestframework==3.15.2
filelock==3.18.0
fsspec==2025.3.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.1.0
requests==2.32.3
setuptools==76.1.0
sqlparse==0.5.0
sympy==1.13.1
torch==2.6.0
typing_extensions==4.12.2
urllib3==2.3.0
